+++ docs/src/releases/Changelog.rst
@@ -5,7 +5,10 @@ Changelog
 Current Development
 -------------------
 
+* [Added] `YAN-1370 <https://jira.skatelescope.org/browse/YAN-1370>`_ 
 
+  - Introduced a new endpoint: /ingestnewmetadata (POST). This endpoint allows you to ingest data product metadata directly through the REST API. Send a POST request with the contents of your metadata file formatted as JSON. The API will parse the JSON data and add the corresponding data product to the metadata store in use.
+ 
 
 Released
 ========